+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Change log
 All notable changes to this project will be documented in this file.
 
+## [0.1.2] - [2018-12-16]
+
+### Added
+
+- Default width and height
+- Error log when something fails while creating images
+
+### Fixed
+
+- Make image size match block attributes width
+
+## [0.1.1] - [2018-12-15]
+
+### Fixed
+
+- Improved resolution based on `images` and `imagesoutdir` configuration entries
+
 ## [0.1.0] - [2018-12-15]
 
 First release
@@ -10,8 +27,4 @@ First release
 - Extension
 - Block processor
 
-## [0.1.1] - [2018-12-15]
 
-### Fix
-
-- Improved resolution based on `images` and `imagesoutdir` configuration entries